ingestion/requirements.txt
boto3
pandas
faker
python-dotenv
